[breaking] runScript*() now returns execution details
The return value from runScript*() was just a string containing stdout
capture (or, a tuple also containing the script globals namespace dict)
Changed return value of runScript*() functions so it's always the same
type. Specifically, it now returns a named tuple also containing stderr,
input() prompts, and exit code.

The default behavior of not returning the namespace remains.
Capturing input() prompts (as a sequence of strings) is a new feature.
Capturing stderr is also new, and it can be optionally turned off.