Add support for execution tracing, using snoop
When test failures or errors occur, an execution trace is added to
the corresponding exception as a note. The actual trace is produced
using the snoop library (which is an optional dependency and installing
it will enable the feature).

Currently, only HomeworkModuleTestCase supports tracing: all imported
HomeworkModuleTestCase.__attrnames__ that are either functions or classes
will be automatically traced (at the moment, there is no option to
select a subset or to change it per test method). For classes, all
methods defined *directly* in the class will be traced (this includes
__init__, if defined). No superclass methods will be traced. Again, this
is not configurable at the moment.

Finally, bumping module version (to 0.1.1).