Truncate TraceLog output to max #lines
This is to avoid "spamming" test output when there is an infinite recursion
(or any other error that causes excessive output). The max number of lines
to keep is currently hardcoded to 500 (which, for infinite recursion still
produces about 1.7Mb of output until stack limit is reached, but that is down
from 20+Mb without truncation...)

Also, bump minor version.